+++ full-moon/src/tokenizer.rs
@@ -304,9 +304,9 @@ impl fmt::Display for Token {
                 quote_type,
             } => {
                 if let Some(blocks) = multi_line {
-                    format!("[{0}[{1}]{0}]", "=".repeat(*blocks), literal.to_string())
+                    format!("[{0}[{1}]{0}]", "=".repeat(*blocks), literal)
                 } else {
-                    format!("{0}{1}{0}", quote_type.to_string(), literal.to_string())
+                    format!("{0}{1}{0}", quote_type.to_string(), literal)
                 }
             }
             Symbol { symbol } => symbol.to_string(),