+++ full-moon/tests/roblox_cases/fail/parser/generic_string/source.lua
@@ -1,0 +1,1 @@
+type Foo = "bar"<number>